+++ internal/mode/validate/doc.go
@@ -1,3 +1,3 @@
-// Package validate is a Ferrite mode that validates the environment variables
-// against the variable definitions and exits the process if any are invalid.
+// Package validate is a Ferrite mode that validates environment variable values
+// against a set of specifications and exits the process if any are invalid.
 package validate